Prod Upload

+++ buildwaiverdata.js
@@ -5,28 +5,28 @@ const dataDir = './_data';
 let waiversFile;
 let oldData;
 let newData;
-const DATAURL = "https://portal-test.forms.gov/mia-test/madeinamericanonavailabilitywaiverrequest/submission?created__gt=2021-10-13&select=state,data.requestStatus,data.psc,data.procurementTitle,data.contractingOfficeAgencyName,data.waiverCoverage, data.contractingOfficeAgencyId,data.fundingAgencyId,data.fundingAgencyName,data.procurementStage,data.naics,data.summaryOfProcurement,data.waiverRationaleSummary,data.sourcesSoughtOrRfiIssued,data.expectedMaximumDurationOfTheRequestedWaiver,data.isPricePreferenceIncluded,created,modified,data.ombDetermination,data.conditionsApplicableToConsistencyDetermination,data.solicitationId"
-const GITHUBURL = "https://api.github.com/repos/GSA/made-in-america-data/contents/waivers-data.json"
+const DATAURL = "https://submission.forms.gov/mia-live/madeinamericanonavailabilitywaiverrequest/submission?&select=state,data.requestStatus,data.psc,data.procurementTitle,data.contractingOfficeAgencyName,data.waiverCoverage, data.contractingOfficeAgencyId,data.fundingAgencyId,data.fundingAgencyName,data.procurementStage,data.naics,data.summaryOfProcurement,data.waiverRationaleSummary,data.sourcesSoughtOrRfiIssued,data.expectedMaximumDurationOfTheRequestedWaiver,data.isPricePreferenceIncluded,created,modified,data.ombDetermination,data.conditionsApplicableToConsistencyDetermination,data.solicitationId"
+const GITHUBURL = "https://api.github.com/repos/GSA/made-in-america-data/contents/waivers-data-prod.json"
 const API_KEY = process.env.GH_API_KEY
 const FORMSKEY = process.env.FORMS_API_KEY;
 
 async function loadData() {
   try {
-    await smokeCheck();
-    await addNewWaivers();
-    await updateReviewedWaivers();
-    await pushtoRepo(oldData)
-    console.log('COMPLETED')
+      smokeCheck();
+      await addNewWaivers();
+      updateReviewedWaivers();
+      pushtoRepo(oldData)
+      console.log('COMPLETED')
 
   } catch (err) {
     console.log(`${err}`);
   }
-
 }
 
 async function getData(url) {
   try {
     console.log('async data request...')
+    // * result is the data from Forms and the token is the API key
     const result = await axios(url, {
       method: 'get',
       headers: {
@@ -34,6 +34,8 @@ async function getData(url) {
       }
     })
     const sha = result.data.sha;
+    // * if the data is encoded, we need to convert it back to utf-8 in
+    // * in order to read the contents and do any type of manipulation
     const mappedData = (ajaxdata) => {
       if(ajaxdata.data.encoding === 'base64') {
         console.log('Converting BASE 64 to UTF-8') 
@@ -41,7 +43,7 @@ async function getData(url) {
         let text = buffObj.toString('utf-8')      
         ajaxdata.data = JSON.parse(text)
       }
-   
+      // * ...string manipulation for better readable text for the front end
       return ajaxdata.data.map(item => {
         let temp = Object.assign({}, item)
         if(temp.data.expectedMaximumDurationOfTheRequestedWaiver === 'between2And3Years') {
@@ -85,6 +87,8 @@ async function getData(url) {
     }
     
    let final = mappedData(result)
+  // * if the sha value isn't undefined, then creating a new key:value pair in the 
+  // * JSON for the sha value
    if(sha) {
      console.log('including sha value...')
      final['sha'] = sha
@@ -97,23 +101,30 @@ async function getData(url) {
   }
 }
 
-async function smokeCheck() {
+async function smokeCheck () {
   try {
     console.log('checking if files exist...')
+    // * if the waivers-data.json doesn't exists, create the file in the directory...
     if (!fs.existsSync(`${dataDir}/waivers-data.json`)) {
+      // * ...assign the waiversFile variable to the file
       waiversFile = `${dataDir}/waivers-data.json`
       console.log('file not here')
+      // * and create and stringify an empty array in the file
       fs.writeFileSync(waiversFile, JSON.stringify([]), (err) => {
         if (err) { console.log('err', err) } else {
           console.log('data written to file')
         }
       })
+      // * make ajax call to Forms DB to get waiver data...
       await getData(DATAURL).then(res => {
+        //  *...write the data to the waivers-data.json
         fs.writeFileSync(waiversFile, JSON.stringify(res), 'utf-8', null, 2)
+        // * ...and assign the oldData variable to the waivers-data.json file
         oldData = JSON.parse(fs.readFileSync(waiversFile, 'utf-8'))
         return;
       })
     } else { 
+      // * But if the file is present, then just assign the oldData variable to the file.
       oldData = JSON.parse(fs.readFileSync(`${dataDir}/waivers-data.json`, 'utf-8'))
       console.log('Smoke Check completed')
     }
@@ -124,18 +135,18 @@ async function smokeCheck() {
 }
 
 async function addNewWaivers() {
-  //if there is no new data in the directory
+  // * if there is no current waivers file in the directory
   if(!fs.existsSync(`${dataDir}/current-waivers.json`)) {
-    //get the data and write it to json
-    await getData(DATAURL).then(res => {
-      console.log('ADDING NEW WAIVERS!!!!!!')
+    // * go get the data from Forms DB...
+     await getData(DATAURL).then(res => {
+      // * and write it to json
       fs.writeFileSync(`${dataDir}/current-waivers.json`, JSON.stringify(res), 'utf-8', null, 2)
-      //and lets call it newData
+      console.log('ADDING NEW WAIVERS!!!!!!')
+      // * and lets call it newData
       newData = JSON.parse(fs.readFileSync(`${dataDir}/current-waivers.json`, 'utf-8'))
-      // update our waiver data file
+      // * filter out the data that does no exist in the old data
       const diff = newData.filter(n => !oldData?.some(item => n._id === item._id))
-      console.log('here2') 
-      //compare objects that are not in old and write them into the new file
+      // * and write them into the new file
       fs.writeFileSync(`${newData}`, JSON.stringify(diff), 'utf-8')
       console.log('FINISHED ADDING NEW WAIVERS...')
       console.log('There are ' + newData.length + ' waivers in the current file')
@@ -146,43 +157,60 @@ async function addNewWaivers() {
 }
 
 
-async function pushtoRepo(data) {
-   await ajaxMethod(data,'')
+function pushtoRepo(data) {
+  /** ajaxMethod
+   * @param data is the current-waviers.json
+   * @param '' is the sha value
+   * * when pushing to the repo when the file isn't present, you don't need a sha value
+   * * but on updates and deletions a sha value is required 
+   */
+   ajaxMethod(data,'')
 }
 
 async function updateRepo(data) {
   console.log('getting SHA Value for Update')
-  let response = await getData(GITHUBURL);
+   const response = await getData(GITHUBURL);
   const shaValue = response.sha;
   ajaxMethod(data,shaValue)
 }
 
-updateReviewedWaivers = () => {
-  // //slice out and replace based on modified date
+function updateReviewedWaivers () {
   console.log('Updating Waivers with new modified date')
-  let modifiedWaivers = compareJSONsforChangesInModifiedDate(oldData, newData)
-  let final = newData.map(obj => modifiedWaivers.find(o => obj._id === o._id) || obj)
-  fs.writeFileSync(`${dataDir}/waivers-data.json`, JSON.stringify(final), 'utf-8')
-  fs.unlinkSync(`${dataDir}/current-waivers.json`)
+  //  * function checks for json waivers that have changed modified data
+  const modifiedWaivers = compareJSONsforChangesInModifiedDate(oldData, newData)
+  // * map the currentdata.json and into a new array and find the objects from the returned 
+  // * 'compareJSONsforChangesInModifiedDate' function
+  if(newData) {
+    console.log('in new data')
+    const final = newData.map(obj => modifiedWaivers.find(o => obj._id === o._id) || obj)
+    // * and replace them.
+    fs.writeFileSync(`${dataDir}/waivers-data.json`, JSON.stringify(final), 'utf-8')
+    // * delete the current waiver file as it's not longer needed till the next pull
+    fs.unlinkSync(`${dataDir}/current-waivers.json`)
+  }
 }
 
 
 function compareJSONsforChangesInModifiedDate(prev, current) {
-   var result = current.filter(({modified}) =>
-    !prev.some(o => new Date(o.modified).getTime() === new Date(modified).getTime())
-  );
-  return result
+    // * return the objects that do not have the same modified date. 
+     const result = current.filter(({modified}) =>
+    //  * ...convert Date object to correctly compare date
+      !prev.some(o => new Date(o.modified).getTime() === new Date(modified).getTime())
+    );
+    return result;
 }
 
 function ajaxMethod(data, shaValue) {
+  // * when pushing to github, the data must be encoded to base64 format
   let buffered = Buffer.from(JSON.stringify(data)).toString('base64') 
-  var jsondata = JSON.stringify({
+  //  * and then the commit message, and all data must be stringfied
+  let jsondata = JSON.stringify({
       "message": "uploading a json file file again",
       "content": buffered,
       "sha" : shaValue
   })
 
-  var config = {
+  let config = {
     method: 'put',
       url: GITHUBURL,
       headers: {
@@ -198,6 +226,12 @@ function ajaxMethod(data, shaValue) {
         return JSON.stringify(response.data);
       })
       .catch(function (error) {
+        /**
+         * ! if there is a 409 error, it means that there is a conflict in that the 
+         * ! file already exists and because did not pass the sha value.
+         * ! In order to update/delete, you must do a GET call to the file and THEN perform 
+         * ! another PUT request
+         */
         if(error.response.status === 409) {
           console.log('409 error!!!!!!!!')
           updateRepo(data);
@@ -208,11 +242,3 @@ function ajaxMethod(data, shaValue) {
 }
 
 loadData()
-
-   // create new hash
-  //  const buffer = crypto.randomBytes(16);
-
-  //  const latest = {
-  //    integrity: buffer.toString('hex'),
-  //    data: [...oldData, ...diff]
-  //  }
